requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.3
flask-wtf==1.2.1
flask-mail==0.9.1
flask-restful==0.3.10
flask-cors==4.0.0
python-dotenv==1.0.0
wtforms==3.1.0
pillow==10.1.0
psycopg2-binary==2.9.9
pymysql==1.1.0
email-validator==2.1.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
requests==2.31.0
numpy==1.26.0
pandas==2.1.0
scikit-learn==1.3.0
opencv-python==4.8.1
pyjwt==2.8.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
tabulate==0.9.0
